chore: address issue #149
